Onboarding: frontier-model step with the full engine picker
The codex-login screen grew into 'Choose your frontier model': the five
engine pills over the same per-engine panels Settings renders, via the
new shared FrontierEnginePicker (FrontierModelPane is now a thin wrapper
over it). The codex login rides inside as the ChatGPT panel; Continue
gates on the active engine being healthy (ChatGPT logged in, or a custom
endpoint that passed Test & Select). The step also moved ahead of
permissions, with the plan crossroads riding directly behind it, and
custom engines skip the crossroads entirely. Motion: tab switches
crossfade with a pinned header (no slide, no re-centering jump).
