+++ Sentient OS macOS/Views/FrontierEnginePicker.swift
@@ -1,0 +1,473 @@
+//
+//  FrontierEnginePicker.swift
+//  Sentient OS macOS
+//
+//  The shared frontier-engine picker — the five engine pills (EngineTab) over each engine's
+//  setup panel, rendered identically by Settings → Frontier Model Choice and onboarding's
+//  choose-your-frontier-model step. Owns the tab state, the endpoint fields
+//  (ModelBackend/CustomProvider — the one source of truth), Test & Select (vision-gated
+//  activation: an unproven model can never become the engine), and the honest local-models
+//  warning. The ChatGPT panel is the host's slot: Settings points at Permissions & Health,
+//  onboarding embeds the live codex login. Layout knob: Settings' fixed 3+2 grid, or
+//  onboarding's single centered row. Deep doc: Documentation/Frontier Model Choice (BYOM).md.
+//
+
+import SwiftUI
+
+struct FrontierEnginePicker<ChatGPTPanel: View>: View {
+
+    /// Settings' fixed two-row grid (3 + 2 — a scrollbar appearing must never reflow the strip,
+    /// the jumpy-rewrap of 2026-07-24), or onboarding's single centered row of five.
+    enum Layout { case settingsGrid, singleRow }
+
+    /// The five engine tabs. OpenRouter, LM Studio, and Custom share the endpoint plumbing;
+    /// LM Studio's tab raises the honest local-models warning on first visit (model
+    /// capability, not plumbing, is the local blocker — the translator solved the plumbing).
+    enum Tab: String, CaseIterable, Identifiable {
+        case chatgpt, claude, openRouter, lmStudio, custom
+        var id: Self { self }
+
+        var label: String {
+            switch self {
+            case .chatgpt:    return "ChatGPT Subscription"
+            case .claude:     return "Claude Subscription"
+            case .openRouter: return "OpenRouter"
+            case .lmStudio:   return "LM Studio"
+            case .custom:     return "Custom"
+            }
+        }
+
+        var badge: String? {
+            switch self {
+            case .chatgpt:  return "recommended"
+            case .claude:   return "coming soon"
+            case .lmStudio: return "local"
+            case .custom:   return "local"
+            case .openRouter: return nil
+            }
+        }
+    }
+
+    /// The one model that cleared the computer-use bar in the 2026-07-24 survey (~10 models,
+    /// real tasks). Prefilled on the OpenRouter tab and named in its note.
+    private static var kimiSlug: String { "moonshotai/kimi-k3" }
+
+    let layout: Layout
+    /// Whether the ChatGPT engine counts as healthy (wears the pill's active dot). Settings
+    /// passes true — login honesty lives in Permissions & Health there; onboarding passes the
+    /// live codex login state.
+    let chatgptHealthy: Bool
+    /// Onboarding greys Test & Select until the background codex install answers — the CLI is
+    /// the engine room even for custom endpoints (the vision probe runs through `codex exec`).
+    /// Settings always passes true.
+    let codexReady: Bool
+    private let chatgptPanel: () -> ChatGPTPanel
+
+    init(layout: Layout = .settingsGrid,
+         chatgptHealthy: Bool = true,
+         codexReady: Bool = true,
+         @ViewBuilder chatgptPanel: @escaping () -> ChatGPTPanel) {
+        self.layout = layout
+        self.chatgptHealthy = chatgptHealthy
+        self.codexReady = codexReady
+        self.chatgptPanel = chatgptPanel
+    }
+
+    @AppStorage(ModelBackend.key) private var backendRaw = ModelBackend.chatgpt.rawValue
+    @AppStorage(CustomProvider.presetKey) private var presetRaw = CustomProvider.Preset.openRouter.rawValue
+    @AppStorage(CustomProvider.baseURLKey) private var baseURL = ""
+    @AppStorage(CustomProvider.modelNameKey) private var modelName = ""
+    @AppStorage(CustomProvider.reasoningKey) private var reasoningRaw = "low"
+    /// Set only by a passing Test Model run; any edit below clears it.
+    @AppStorage(CustomProvider.visionVerifiedKey) private var verified = false
+
+    @State private var tab: Tab = .chatgpt
+    @State private var apiKey = ""
+    @State private var showLocalWarning = false
+    /// The local-reality popup fires once per picker visit, on the LM Studio tab.
+    @State private var localWarningShown = false
+
+    /// The Test Model probe, narrated: nil = untested, "…" = running, ✓/✗ lines otherwise.
+    @State private var testing = false
+    @State private var testVerdict: String?
+
+    private var backend: ModelBackend { ModelBackend(rawValue: backendRaw) ?? .chatgpt }
+    private var savedPreset: CustomProvider.Preset { CustomProvider.Preset(rawValue: presetRaw) ?? .openRouter }
+
+    /// The tab wearing the "active engine" dot.
+    private var activeTab: Tab {
+        guard backend == .custom else { return .chatgpt }
+        switch savedPreset {
+        case .openRouter: return .openRouter
+        case .lmStudio:   return .lmStudio
+        case .custom:     return .custom
+        }
+    }
+
+    var body: some View {
+        VStack(alignment: layout == .singleRow ? .center : .leading, spacing: 26) {
+            tabStrip
+
+            Group {
+                switch tab {
+                case .chatgpt:    chatgptPanel()
+                case .claude:     claudePanel
+                case .openRouter: endpointPanel(for: .openRouter)
+                case .lmStudio:   endpointPanel(for: .lmStudio)
+                case .custom:     endpointPanel(for: .custom)
+                }
+            }
+            // The Settings editorial measure — panels never stretch wider than prose stays
+            // readable, in either host (SettingsPane caps at the same width).
+            .frame(maxWidth: 640, alignment: .leading)
+            .id(tab)
+            // A quiet crossfade — no lateral slide: panels differ in height, and a slide on
+            // top of the height change read as an abrupt jump (field feedback 2026-07-25).
+            .transition(.opacity)
+        }
+        .animation(.easeInOut(duration: 0.25), value: tab)
+        .onAppear {
+            tab = activeTab
+            apiKey = CustomProvider.apiKey ?? ""
+        }
+        // The host's ChatGPT panel flipping the backend home (Use ChatGPT) retires any verdict.
+        .onChange(of: backendRaw) {
+            if backend == .chatgpt { testVerdict = nil }
+        }
+        .alert("A note on local frontier models", isPresented: $showLocalWarning) {
+            Button("Understood") {}
+        } message: {
+            Text("Most models you can run locally are far too weak for computer use; in our testing only Kimi K3 performed well enough, and most other open-weights models aren't multimodal (can't see), which computer use requires.\n\nIf you can somehow run Kimi K3 locally, this preset is all yours.\n\nOtherwise run it through OpenRouter, or if privacy is the concern, both OpenAI and Claude subscriptions let you turn off training on your data in your account settings on their own site. Those frontier options remain the best way to use Sentient.")
+        }
+    }
+
+    // MARK: - The engine tab strip
+
+    private var tabStrip: some View {
+        Group {
+            switch layout {
+            case .settingsGrid:
+                VStack(alignment: .leading, spacing: 8) {
+                    HStack(spacing: 8) {
+                        pill(.chatgpt); pill(.claude); pill(.openRouter)
+                    }
+                    HStack(spacing: 8) {
+                        pill(.lmStudio); pill(.custom)
+                    }
+                }
+            case .singleRow:
+                HStack(spacing: 8) {
+                    ForEach(Tab.allCases) { pill($0) }
+                }
+            }
+        }
+        .fixedSize()
+    }
+
+    private func pill(_ t: Tab) -> some View {
+        EngineTab(label: t.label, badge: t.badge,
+                  selected: tab == t, active: activeTab == t && isActiveEngineHealthy(t)) {
+            selectTab(t)
+        }
+    }
+
+    /// The active dot stays honest: a custom engine only wears it once it's proven usable,
+    /// and ChatGPT's honesty is the host's call (Settings: always; onboarding: logged in).
+    private func isActiveEngineHealthy(_ t: Tab) -> Bool {
+        t == .chatgpt ? chatgptHealthy : CustomProvider.current.isUsable
+    }
+
+    /// Browsing tabs never disturbs a live engine: fields are only prefilled while nothing
+    /// custom is active (Test Model and Use-this-model pin the right values when the user
+    /// actually acts). The LM Studio tab raises the local-reality popup once per visit.
+    private func selectTab(_ t: Tab) {
+        tab = t
+        testVerdict = nil
+        if t == .lmStudio, !localWarningShown {
+            localWarningShown = true
+            showLocalWarning = true
+        }
+        guard backend != .custom else { return }
+        switch t {
+        case .openRouter:
+            if modelName.isEmpty { modelName = Self.kimiSlug }   // Kimi K3 forward
+            baseURL = CustomProvider.Preset.openRouter.defaultBaseURL
+        case .lmStudio:
+            if baseURL.isEmpty || baseURL == CustomProvider.Preset.openRouter.defaultBaseURL {
+                baseURL = CustomProvider.Preset.lmStudio.defaultBaseURL
+            }
+        default:
+            break
+        }
+    }
+
+    // MARK: - Claude (coming soon)
+
+    private var claudePanel: some View {
+        SettingsGroup(label: "Your Claude Plan", badge: "coming soon") {
+            VStack(alignment: .leading, spacing: 10) {
+                SettingsProse("Your Claude subscription will power Sentient the same way ChatGPT does today, through Claude Code's own command line. It's on the bench and coming soon.")
+                MonoCaps("claude -p · in the works", size: 8.5, tracking: 1.6, color: Theme.Ink.deepMuted)
+            }
+        }
+    }
+
+    // MARK: - Configurable endpoints (OpenRouter · Custom + its local presets)
+
+    private func endpointPanel(for preset: CustomProvider.Preset) -> some View {
+        SettingsGroup(label: preset == .openRouter ? "Your OpenRouter"
+                           : preset == .lmStudio ? "Your LM Studio" : "Your Endpoint") {
+            VStack(alignment: .leading, spacing: 14) {
+                switch preset {
+                case .openRouter:
+                    SettingsProse("Any model on OpenRouter, billed to your own key.")
+                case .lmStudio:
+                    SettingsProse("A model running on your own hardware, through LM Studio's local server. Sentient's translator makes it a first-class engine, computer use included; whether the model is up to the job is the honest question (see the note when this tab opens).")
+                case .custom:
+                    SettingsProse("Any endpoint that speaks the OpenAI Responses API (a /v1/responses route). Base URL, model name, key if it needs one.")
+                }
+                if preset != .lmStudio {
+                    SettingsProse("One important note: of every model we tested, Kimi K3 at low reasoning is the only open-weights model that can reliably drive computer use. Claude Sonnet 5 with reasoning off, and GPT-5.6 Sol at low reasoning are also great options we can stand by.")
+                }
+
+                // OpenRouter's base URL is fixed — no field, the tab pins it itself.
+                if preset != .openRouter {
+                    fieldRow(label: "BASE URL",
+                             placeholder: preset == .lmStudio
+                                 ? CustomProvider.Preset.lmStudio.defaultBaseURL
+                                 : "https://your-endpoint.example/v1",
+                             text: $baseURL)
+                }
+                fieldRow(label: "MODEL",
+                         placeholder: preset == .openRouter ? Self.kimiSlug
+                             : preset == .lmStudio ? "the model id shown in LM Studio"
+                             : "the model id your server expects",
+                         text: $modelName)
+                secureRow(label: preset == .openRouter ? "OPENROUTER API KEY" : "API KEY",
+                          placeholder: preset == .openRouter ? "sk-or-…" : "optional for keyless servers")
+
+                reasoningField
+
+                SettingsProse("Your model has to be able to see: Sentient acts on your Mac by looking at the screen, so a text-only model can't drive it. The test below checks that for you by asking your model to read a picture.")
+
+                HStack(spacing: 10) {
+                    SettingsPillButton(title: testing ? "Testing…" : "Test & Select",
+                                       tint: codexReady ? Theme.Ink.bright : Theme.Ink.deepMuted) {
+                        guard codexReady, !testing else { return }
+                        runTest(preset: preset)
+                    }
+                    if isActive(preset) {
+                        FrontierActiveLine("This model is running Sentient.")
+                    }
+                }
+                if !codexReady {
+                    MonoWaitLine("installing codex in the background…")
+                }
+                if let testVerdict {
+                    SettingsProse(testVerdict)
+                }
+
+                SettingsHairline()
+                SettingsProse("Gmail and Calendar ride your ChatGPT account, so they sit out while a custom model is active. Everything else works, and your data still never leaves this Mac except what the model itself is sent.")
+            }
+        }
+    }
+
+    /// The endpoint's ONE reasoning level — free text, because models speak different dialects
+    /// (`none`, `low`, `xhigh`, `adaptive`, …), applied to every run this model powers (the
+    /// Speed slider is ChatGPT's). Kimi wants low; Claude-class endpoints need none. Editing it
+    /// re-runs the gate via the field's invalidate, since a wrong level can break the endpoint
+    /// outright.
+    private var reasoningField: some View {
+        VStack(alignment: .leading, spacing: 6) {
+            fieldRow(label: "REASONING",
+                     placeholder: "low · none · xhigh · whatever your model supports",
+                     text: $reasoningRaw)
+            MonoCaps("applies everywhere this model runs", size: 7.5, tracking: 1.6,
+                     color: Theme.Ink.deepMuted)
+        }
+    }
+
+    // MARK: - Pieces
+
+    private func fieldRow(label: String, placeholder: String, text: Binding<String>) -> some View {
+        VStack(alignment: .leading, spacing: 6) {
+            MonoCaps(label, size: 8.5, tracking: 2.0, color: Theme.Ink.deepMuted)
+            TextField(placeholder, text: text)
+                .textFieldStyle(.plain)
+                .font(.system(size: 11.5)).foregroundStyle(Theme.Ink.statusInk)
+                .padding(.horizontal, 12).padding(.vertical, 8)
+                .background(Color.white.opacity(0.02), in: RoundedRectangle(cornerRadius: 9, style: .continuous))
+                .overlay(RoundedRectangle(cornerRadius: 9, style: .continuous)
+                    .strokeBorder(Theme.stroke, lineWidth: 1))
+                .onChange(of: text.wrappedValue) { invalidate() }
+        }
+    }
+
+    /// Any edit to the endpoint retires the old verdict: a different model has to prove it can
+    /// see for itself. Also drops the backend back to ChatGPT if the running engine just became
+    /// unverified, so Sentient is never left pointed at an unproven model.
+    private func invalidate() {
+        guard verified || backend == .custom else { return }
+        verified = false
+        testVerdict = nil
+        if backend == .custom { backendRaw = ModelBackend.chatgpt.rawValue }
+    }
+
+    private func secureRow(label: String, placeholder: String) -> some View {
+        VStack(alignment: .leading, spacing: 6) {
+            MonoCaps(label, size: 8.5, tracking: 2.0, color: Theme.Ink.deepMuted)
+            SecureField(placeholder, text: $apiKey)
+                .textFieldStyle(.plain)
+                .font(.system(size: 11.5)).foregroundStyle(Theme.Ink.statusInk)
+                .padding(.horizontal, 12).padding(.vertical, 8)
+                .background(Color.white.opacity(0.02), in: RoundedRectangle(cornerRadius: 9, style: .continuous))
+                .overlay(RoundedRectangle(cornerRadius: 9, style: .continuous)
+                    .strokeBorder(Theme.stroke, lineWidth: 1))
+                .onChange(of: apiKey) {
+                    let trimmed = apiKey.trimmingCharacters(in: .whitespacesAndNewlines)
+                    if trimmed.isEmpty {
+                        Keychain.delete(CustomProvider.apiKeyAccount)
+                    } else {
+                        Keychain.set(CustomProvider.apiKeyAccount, trimmed)
+                    }
+                    invalidate()
+                }
+            MonoCaps("stored in your mac's keychain", size: 7.5, tracking: 1.6, color: Theme.Ink.deepMuted)
+        }
+    }
+
+    private func isActive(_ preset: CustomProvider.Preset) -> Bool {
+        backend == .custom && savedPreset == preset && CustomProvider.current.isUsable
+    }
+
+    /// Flip the backend to the viewed tab's endpoint — reached ONLY through a passing
+    /// Test & Select run, so an unproven model can never become the engine. Fields already
+    /// autosave (@AppStorage); activation is just the choice becoming official.
+    private func activate(_ preset: CustomProvider.Preset) {
+        presetRaw = preset.rawValue
+        guard CustomProvider.current.isUsable else { return }
+        backendRaw = ModelBackend.custom.rawValue
+    }
+
+    private func runTest(preset: CustomProvider.Preset) {
+        // OpenRouter's URL is pinned (no field on that tab); elsewhere only fill an empty one.
+        if preset == .openRouter {
+            baseURL = preset.defaultBaseURL
+        } else if baseURL.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
+            baseURL = preset.defaultBaseURL
+        }
+        presetRaw = preset.rawValue   // the probe reads the SAVED provider — keep it current
+        guard CustomProvider.current.isConfigured else {
+            testVerdict = "✗ It needs a base URL and a model name first."
+            return
+        }
+        testing = true
+        testVerdict = "Showing your model a picture and asking it to read the number… local models can take a minute to load."
+        Task {
+            let verdict = await CodexCLI.probeCustomEndpoint()
+            await MainActor.run {
+                testing = false
+                switch verdict {
+                case .available:
+                    verified = true
+                    activate(preset)   // Test & Select: a passing model becomes the engine
+                    testVerdict = "✓ Your model answered and read the picture. Sentient is now running on it."
+                case .notInstalled:
+                    verified = false
+                    testVerdict = "✗ Codex CLI is missing on this Mac. Install it in Permissions & Health first; it's the engine room even for custom models."
+                case .notWorking(let detail):
+                    verified = false
+                    testVerdict = "✗ \(friendly(detail))"
+                }
+            }
+        }
+    }
+
+    /// Reduce codex's raw stderr to one readable hint; the full detail stays in the console.
+    /// A base URL missing its `/v1` is the single most common setup slip (codex appends
+    /// `/responses` to whatever you give it, so the request lands on a route the server
+    /// doesn't serve), so every failure carries that nudge when it applies.
+    private func friendly(_ detail: String) -> String {
+        let lowered = detail.lowercased()
+        let trimmedURL = baseURL.trimmingCharacters(in: .whitespacesAndNewlines)
+            .trimmingCharacters(in: CharacterSet(charactersIn: "/"))
+        let v1Hint = trimmedURL.hasSuffix("/v1") ? ""
+            : " Most servers also expect the base URL to end in /v1."
+        if lowered.hasPrefix("blind") || lowered.contains("image input") || lowered.contains("multimodal") {
+            return "This model can't see pictures, so it can't act on your Mac. Pick one that accepts images."
+        }
+        if lowered.contains("401") || lowered.contains("unauthorized") {
+            return "The endpoint rejected the API key."
+        }
+        if lowered.contains("connection refused") || lowered.contains("error sending request") {
+            return "Nothing is listening at that base URL. Is the server running?\(v1Hint)"
+        }
+        if lowered.contains("404") || lowered.contains("not found")
+            || lowered.contains("unexpected endpoint") {
+            return "The endpoint answered but has no /v1/responses route.\(v1Hint.isEmpty ? " It must support the OpenAI Responses API." : v1Hint)"
+        }
+        return "No answer from the endpoint.\(v1Hint) \(detail.prefix(180))"
+    }
+}
+
+// MARK: - The engine tab
+
+/// One tab in the engine strip: a fixed-size pill that reads like a place, not a button —
+/// EVERY pill the same width and height so the row's geometry never shifts (scrollbars,
+/// badges, and selection can't reflow it). Selected = elevated wash + white ring; the ACTIVE
+/// engine wears the small green status dot (the same honest LED the health rows use). Badges
+/// whisper in mono-caps under the label.
+struct EngineTab: View {
+    static var pillWidth: CGFloat { 176 }
+    static var pillHeight: CGFloat { 52 }
+
+    let label: String
+    var badge: String? = nil
+    let selected: Bool
+    let active: Bool
+    let action: () -> Void
+
+    var body: some View {
+        Button(action: action) {
+            VStack(alignment: .leading, spacing: 3) {
+                HStack(spacing: 7) {
+                    if active { HealthDot(color: Theme.Ink.green) }
+                    Text(label)
+                        .font(.system(size: 12.5, weight: selected ? .medium : .regular))
+                        .foregroundStyle(selected ? .white : Theme.Ink.body)
+                        .lineLimit(1)
+                        .minimumScaleFactor(0.92)
+                }
+                if let badge {
+                    MonoCaps(badge, size: 7, tracking: 1.4,
+                             color: badge == "recommended" ? Theme.Ink.green.opacity(0.85)
+                                                           : .white.opacity(0.45))
+                }
+            }
+            .frame(maxWidth: .infinity, alignment: .leading)
+            .padding(.horizontal, 12)
+            .frame(width: Self.pillWidth, height: Self.pillHeight)
+            .background(selected ? Theme.elevated : Color.white.opacity(0.02),
+                        in: RoundedRectangle(cornerRadius: 14, style: .continuous))
+            .overlay(RoundedRectangle(cornerRadius: 14, style: .continuous).strokeBorder(
+                selected ? Color.white.opacity(0.28) : Color.white.opacity(0.10), lineWidth: 1))
+            .contentShape(RoundedRectangle(cornerRadius: 14, style: .continuous))
+        }
+        .buttonStyle(PressScaleStyle())
+    }
+}
+
+/// The green "this engine is live" line — shared by the picker's endpoint panels and the
+/// Settings pane's ChatGPT panel.
+struct FrontierActiveLine: View {
+    let text: String
+    init(_ text: String) { self.text = text }
+
+    var body: some View {
+        HStack(spacing: 8) {
+            HealthDot(color: Theme.Ink.green)
+            Text(text).font(.system(size: 12, weight: .medium)).foregroundStyle(.white)
+        }
+    }
+}
